Moved tests out of integration/test_tpp into unit/test_tables
